Separet function to helper and add set state history to run dropdown
